Restore accidental removal of HapticClick logic
While bringing up the long squeeze I have accidentally
removed the vibration logic for actions that can't be
triggered while the screen is off.

With this version of the logic, inside onProgress, we
only check for the single action since the long squeeze
action is strictly connected to the single one.

Change-Id: Ic98250cc36c1df9e78e10128d531713ffb59f257